requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.3
django-adminlte3==0.1.6
django-phone-field==1.8.1
djangorestframework==3.14.0
flake8==6.0.0
gunicorn==21.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
